Moved storage utility functions to core module
Some storage functions are generic enough that they can be shared in the core module, so that they aren't duplicated between the browser and node modules.
This moves the oidc context read/write function from the node to the core module, ensures 100% code coverage on the moved code, and moves the tests alongside the code in the core.
