+++ packages/core/src/index.ts
@@ -57,7 +57,12 @@ export { default as IStorage } from "./storage/IStorage";
 export { default as validateSchema } from "./util/validateSchema";
 
 export { default as IStorageUtility } from "./storage/IStorageUtility";
-export { default as StorageUtility } from "./storage/StorageUtility";
+export {
+  default as StorageUtility,
+  OidcContext,
+  loadOidcContextFromStorage,
+  saveSessionInfoToStorage,
+} from "./storage/StorageUtility";
 export { default as InMemoryStorage } from "./storage/InMemoryStorage";
 
 export { default as ConfigurationError } from "./errors/ConfigurationError";
