+++ packages/node/src/login/oidc/redirectHandler/AuthCodeRedirectHandler.spec.ts
@@ -419,129 +419,6 @@ describe("AuthCodeRedirectHandler", () => {
       ).rejects.toThrow("The IdP did not return the expected tokens.");
     });
 
-    it("throws if no stored state matches the current request's", async () => {
-      const mockedStorage = mockStorageUtility({
-        "solidClientAuthenticationUser:mySession": {
-          issuer: "https://my.idp/",
-          codeVerifier: "some code verifier",
-          redirectUri: "https://my.app/redirect",
-          dpop: "true",
-        },
-      });
-
-      const authCodeRedirectHandler = getAuthCodeRedirectHandler({
-        storageUtility: mockedStorage,
-      });
-
-      await expect(
-        authCodeRedirectHandler.handle(
-          "https://my.app/redirect?code=someCode&state=someState"
-        )
-      ).rejects.toThrow(
-        "Failed to retrieve OIDC context from storage for login request associated with state"
-      );
-    });
-
-    it("throws if no issuer is stored for the user", async () => {
-      const mockedStorage = mockStorageUtility({
-        "solidClientAuthenticationUser:someState": {
-          sessionId: "mySession",
-        },
-        "solidClientAuthenticationUser:mySession": {
-          codeVerifier: "some code verifier",
-          redirectUri: "https://my.app/redirect",
-          dpop: "true",
-        },
-      });
-
-      const authCodeRedirectHandler = getAuthCodeRedirectHandler({
-        storageUtility: mockedStorage,
-      });
-
-      await expect(
-        authCodeRedirectHandler.handle(
-          "https://my.app/redirect?code=someCode&state=someState"
-        )
-      ).rejects.toThrow(
-        "Failed to retrieve OIDC context from storage for login request associated with state"
-      );
-    });
-
-    it("throws if no code verifier is stored for the user", async () => {
-      const mockedStorage = mockStorageUtility({
-        "solidClientAuthenticationUser:someState": {
-          sessionId: "mySession",
-        },
-        "solidClientAuthenticationUser:mySession": {
-          issuer: "https://my.idp/",
-          redirectUri: "https://my.app/redirect",
-          dpop: "true",
-        },
-      });
-
-      const authCodeRedirectHandler = getAuthCodeRedirectHandler({
-        storageUtility: mockedStorage,
-      });
-
-      await expect(
-        authCodeRedirectHandler.handle(
-          "https://my.app/redirect?code=someCode&state=someState"
-        )
-      ).rejects.toThrow(
-        "Failed to retrieve OIDC context from storage for login request associated with state"
-      );
-    });
-
-    it("throws if no redirect uri is stored for the user", async () => {
-      const mockedStorage = mockStorageUtility({
-        "solidClientAuthenticationUser:someState": {
-          sessionId: "mySession",
-        },
-        "solidClientAuthenticationUser:mySession": {
-          issuer: "https://my.idp/",
-          codeVerifier: "some code verifier",
-          dpop: "true",
-        },
-      });
-
-      const authCodeRedirectHandler = getAuthCodeRedirectHandler({
-        storageUtility: mockedStorage,
-      });
-
-      await expect(
-        authCodeRedirectHandler.handle(
-          "https://my.app/redirect?code=someCode&state=someState"
-        )
-      ).rejects.toThrow(
-        "Failed to retrieve OIDC context from storage for login request associated with state"
-      );
-    });
-
-    it("throws if no token type is stored for the user", async () => {
-      const mockedStorage = mockStorageUtility({
-        "solidClientAuthenticationUser:someState": {
-          sessionId: "mySession",
-        },
-        "solidClientAuthenticationUser:mySession": {
-          issuer: "https://my.idp/",
-          codeVerifier: "some code verifier",
-          redirectUri: "https://my.app/redirect",
-        },
-      });
-
-      const authCodeRedirectHandler = getAuthCodeRedirectHandler({
-        storageUtility: mockedStorage,
-      });
-
-      await expect(
-        authCodeRedirectHandler.handle(
-          "https://my.app/redirect?code=someCode&state=someState"
-        )
-      ).rejects.toThrow(
-        "Failed to retrieve OIDC context from storage for login request associated with state"
-      );
-    });
-
     it("throws if the Session manager cannot retrieve the session info", async () => {
       // Sets up the mock-up for token exchange
       const { Issuer } = jest.requireMock("openid-client");
