+++ packages/node/src/login/oidc/redirectHandler/AuthCodeRedirectHandler.ts
@@ -28,12 +28,13 @@ import { inject, injectable } from "tsyringe";
 import {
   IClient,
   IClientRegistrar,
-  IIssuerConfig,
   IIssuerConfigFetcher,
   IRedirectHandler,
   ISessionInfo,
   ISessionInfoManager,
   IStorageUtility,
+  loadOidcContextFromStorage,
+  saveSessionInfoToStorage,
 } from "@inrupt/solid-client-authn-core";
 import { URL } from "url";
 import { IdTokenClaims, Issuer, TokenSet } from "openid-client";
@@ -44,99 +45,6 @@ import {
   buildDpopFetch,
   fetchType,
 } from "../../../authenticatedFetch/fetchFactory";
-
-type OidcContext = {
-  sessionId: string;
-  issuerConfig: IIssuerConfig;
-  codeVerifier: string;
-  redirectUri: string;
-  dpop: boolean;
-};
-
-/**
- * Based on the provided state, this looks up contextual information stored
- * before redirecting the user to the OIDC issuer.
- * @param oauthState The state (~ correlation ID) of the OIDC request
- * @param storageUtility
- * @param configFetcher
- * @returns Information stored about the client issuing the request
- */
-async function loadOidcContextFromStorage(
-  oauthState: string,
-  storageUtility: IStorageUtility,
-  configFetcher: IIssuerConfigFetcher
-): Promise<OidcContext> {
-  try {
-    // Since we throw if not found, the type assertion are ok too
-    const storedSessionId = (await storageUtility.getForUser(
-      oauthState,
-      "sessionId",
-      {
-        errorIfNull: true,
-      }
-    )) as string;
-
-    const [
-      issuerIri,
-      codeVerifier,
-      storedRedirectIri,
-      dpop,
-    ] = (await Promise.all([
-      storageUtility.getForUser(storedSessionId, "issuer", {
-        errorIfNull: true,
-      }),
-      storageUtility.getForUser(storedSessionId, "codeVerifier", {
-        errorIfNull: true,
-      }),
-      storageUtility.getForUser(storedSessionId, "redirectUri", {
-        errorIfNull: true,
-      }),
-      storageUtility.getForUser(storedSessionId, "dpop", { errorIfNull: true }),
-    ])) as string[];
-
-    // Unlike openid-client, this looks up the configuration from storage
-    const issuerConfig = await configFetcher.fetchConfig(issuerIri);
-    return {
-      sessionId: storedSessionId,
-      codeVerifier,
-      redirectUri: storedRedirectIri,
-      issuerConfig,
-      dpop: dpop === "true",
-    };
-  } catch (e) {
-    throw new Error(
-      `Failed to retrieve OIDC context from storage for login request associated with state [${oauthState}]: ${e.toString()}`
-    );
-  }
-}
-
-async function saveSessionInfoToStorage(
-  storageUtility: IStorageUtility,
-  sessionId: string,
-  idToken: string,
-  webId: string,
-  isLoggedIn: string,
-  refreshToken?: string
-): Promise<void> {
-  if (refreshToken !== undefined) {
-    await storageUtility.setForUser(
-      sessionId,
-      {
-        refreshToken,
-      },
-      { secure: true }
-    );
-  }
-  await storageUtility.setForUser(
-    sessionId,
-    {
-      idToken,
-      webId,
-      isLoggedIn,
-    },
-    { secure: true }
-  );
-}
 
 /**
  * Extract a WebID from an ID token payload. Note that this does not yet implement the
@@ -257,7 +165,6 @@ export class AuthCodeRedirectHandler implements IRedirectHandler {
       authFetch = await buildDpopFetch(
         tokenSet.access_token,
         tokenSet.refresh_token,
-        // TODO: buildDpopFetch isn't implemented yet
         // TS thinks dpopKey isn't initialized, when it is.
         // eslint-disable-next-line @typescript-eslint/ban-ts-comment
         // @ts-ignore
